# doc/api.yaml
openapi: '3.0.2'
info:
  title: WASAPhoto Backend API
  version: '1.0'
  description: |-
    The WASAPhoto Backend API allows to perform the following operations:
    - Log in by specifying a username.
    - Get the a stream of photos (images) in reverse chronological order for the logged-in user,
      with information about when each photo was uploaded (date and time) and how many likes
      and comments it has. The stream is composed by photos from “following”
      (other users that the user follows).
    - Place (or  remove) a “like” to photos from other users.
    - Add comments to any image (even those uploaded by themself), and remove
      them (only authors can remove their comments).
    - Ban other users and remove the ban.
    - Get users profiles showing: the user’s photos (in reverse chronological order),
      how many photos have been uploaded, and the user’s followers and following.
    - Change username.
    - Upload photos and remove photos (removal of an image will also remove likes and comments).
    - Follow/unfollow other users.
    - Search other user profiles via username.
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
...
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user 
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody: { $ref: "#/components/requestBodies/DoLogin" }
      security: []
      responses:
        '201': { $ref: "#/components/responses/SuccessfulDoLogin" }
        '400': { $ref: "#/components/responses/BadRequest" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
  /photos/{photoID}:
    parameters:
      - $ref: "#/components/parameters/PathPhotoID"
    get:
      tags: ["photos"]
      summary: Gets a photo
      description: |-
        If the logged-in user is not banned by the author of the photo,
        it returns the requested photo; otherwise it returns an error.
      operationId: getPhoto
      responses:
        '200': { $ref: "#/components/responses/SuccessfulGetPhoto" }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/NotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
    delete:
      tags: ["photos"]
      summary: Deletes a photo
      description: |-
        If the logged-in user is the author of the photo
        with `photoID` given in path, it deletes the photo;
        otherwise, it returns an error.
      operationId: deletePhoto
      responses:
        '200': { $ref: "#/components/responses/SuccessfulDeletePhoto" }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/NotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
  /photos/{photoID}/comments/:
    parameters:
      - $ref: "#/components/parameters/PathPhotoID"
    get:
      parameters:
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/StartId"
      tags: ["comments"]
      summary: Gets the comments for a photo
      description: |-
        If the logged-in user is not banned by the author of the photo, it returns
        the list of comments of the photo; otherwise it returns an error.
      operationId: getComments
      responses:
        '200': { $ref: "#/components/responses/SuccessfulGetComments" }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/NotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
    post:
      tags: ["comments"]
      summary: Adds a comment to a photo
      description: |-
        If the logged-in user is not banned by the author of the photo, it adds
        a comment to the photo authored by the logged-in user;
        otherwise it returns an error.
      operationId: commentPhoto
      requestBody: { $ref: "#/components/requestBodies/CommentPhoto" }
      responses:
        '201': { $ref: "#/components/responses/SuccessfulCommentPhoto" }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/NotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
  /users/{userID}/followers/:
    parameters:
      - $ref: "#/components/parameters/PathUserID"
    get:
      parameters:
        - $ref: "#/components/parameters/StartId"
      tags: ["users"]
      summary: Gets the followers of a user
      description: |-
        If the logged-in user is not banned by the user with `userID` given in path,
        it returns the list of its followers; otherwise it returns an error.
      operationId: getFollowers
      responses:
        '200': { $ref: "#/components/responses/SuccessfulGetFollowers" }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/NotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
  /following/{userID}:
    parameters:
       - $ref: "#/components/parameters/PathUserID"
    put:
      tags: ["following"]
      summary: Adds a user to following
      description: |-
        If the logged-in user is not banned by the user with `userID` given in path,
        it adds him/her to the following of the logged-in user;
        otherwise it returns an error.
      operationId: followUser
      responses:
        '204': { $ref: "#/components/responses/SuccessfulFollowUser" }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/NotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
    delete:
      tags: ["following"]
      summary: Removes a user from following
      description: |-
        It unfollows the user with `userID` given in path
        (or does nothing if it is already not followed);
        otherwise it returns an error.
      operationId: unfollowUser
      responses:
        '204': { $ref: "#/components/responses/SuccessfulUnfollowUser" }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/NotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
  /bans/{userID}:
    parameters:
       - $ref: "#/components/parameters/PathUserID"
    put:
      tags: ["bans"]
      summary: Adds a user to bans
      description: |-
        It adds the user with `userID` given in path to the bans of the logged-in user.
      operationId: banUser
      responses:
        '204': { $ref: "#/components/responses/SuccessfulBanUser" }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/NotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
    delete:
      tags: ["bans"]
      summary: Removes a user from bans
      description: |-
        It removes the user with `userID` given in path from the bans of the logged-in user.
      operationId: unbanUser
      responses:
        '204': { $ref: "#/components/responses/SuccessfulUnbanUser" }
        '400': { $ref: "#/components/responses/BadRequest" }
        '401': { $ref: "#/components/responses/Unauthorized" }
        '403': { $ref: "#/components/responses/Forbidden" }
        '404': { $ref: "#/components/responses/NotFound" }
        '500': { $ref: "#/components/responses/InternalServerError" }
...
components:
  schemas:
...
    dateTime:
      description: A date-time.
      type: string
      format: date-time
      pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
      minLength: 20
      maxLength: 20
    text:
      description: A generic text that does not exceed 2200 characters.
      type: string
      example: lorem ipsum dolor sit amet.
      pattern: '.*'
      minLength: 0
      maxLength: 2200
    photo:
      description: A png photo of 10 MB maximum size in binary format.
      type: string
      format: binary
      pattern: '^[01]+$'
      minLength: 536
      maxLength: 10485760 # 10MB
    username:
      description: Username associated with a unique user.
      type: string
      example: Maria_03
      pattern: '^[a-zA-Z0-9_-]{3,16}$' 
      minLength: 3
      maxLength: 16
    userID:
      description: UUID associated with a unique user.
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      minLength: 36
      maxLength: 36
    photoID:
      description: UUID associated with a unique photo.
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      minLength: 36
      maxLength: 36
    commentId:
      description: UUID associated with a unique comment.
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      minLength: 36
      maxLength: 36
    # Objects
    Paginator:
      description: A generic paginated list.
      type: object
      properties:
        lastDate:
          description: |-
            The date of the last record in the current response.
          type: string
          format: date-time
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
          minLength: 20
          maxLength: 20
        lastId:
          description: |-
            The id of the last record in the current response.
          type: string
          format: uuid
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          minLength: 36
          maxLength: 36
    User:
      description: An existing user.
      type: object
      properties:
        userID: { $ref: "#/components/schemas/userID" }
        username: { $ref: "#/components/schemas/username" }
    UserList:
      description: A list of users.
      allOf:
        - $ref: "#/components/schemas/Paginator"
        - type: object
          properties:
            records:
              description: An array of users.
              type: array
              items: { $ref: "#/components/schemas/User" }
              minItems: 0
              maxItems: 20
...
    Comment:
      description: A comment.
      type: object
      properties:
        commentId: { $ref: "#/components/schemas/commentId" }
        text: { $ref: "#/components/schemas/text" }
        author: { $ref: "#/components/schemas/User" }
        dateTime: { $ref: "#/components/schemas/dateTime" }
    CommentList:
      description: A list of comments.
      allOf:
        - $ref: "#/components/schemas/Paginator"
        - type: object
          properties:
            records:
              description: An array of comments.
              type: array
              items: { $ref: "#/components/schemas/Comment" }
              minItems: 0
              maxItems: 20
  requestBodies:
    DoLogin:
      description: Request body for doLogin operation.
      content:
        application/json:
          schema:
            description: A username (new or existing).
            type: object
            properties:
              username: { $ref: "#/components/schemas/username" }
      required: true
...
    CommentPhoto:
      description: Request body for commentPhoto operation.
      content:
        application/json:
          schema:
            description: |-
              Comment to add to the photo. The comment must not exceed 2200 characters.
            type: object
            properties:
              text: { $ref: "#/components/schemas/text" }
      required: true
  responses:
    InternalServerError:
      description: Unexpected condition was encountered.
    BadRequest:
      description: |-
        The server cannot or will not process the request due to an apparent client error.
    Unauthorized:
      description: Authorization identifier is missing or invalid.
    Forbidden:
      description: logged-in user cannot perform the requested operation.
    NotFound:
      description: Requested resource was not found.
...
    SuccessfulDoLogin:
      description: User log-in action successful.
      content:
        application/json:
          schema:
            description: The requested user UUID.
            type: object
            properties:
              userID: { $ref: "#/components/schemas/userID" }
...
    SuccessfulGetPhoto:
      description: The requested photo.
      content:
        image/png:
          schema: { $ref: "#/components/schemas/photo" }
...
    SuccessfulDeletePhoto:
      description: Photo was deleted successfully.
      content:
        application/json:
          schema:
            description: The UUID of the deleted photo.
            type: object
            properties:
              photoID: { $ref: "#/components/schemas/photoID" }
...
    SuccessfulGetComments:
      description: A list of comments for the photo.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/CommentList" }
    SuccessfulCommentPhoto:
      description: A comment was successfully added to the photo.
      content:
        application/json:
          schema:
            description: The UUID of the newly added comment.
            type: object
            properties:
              commentId: { $ref: "#/components/schemas/commentId" }
...
    SuccessfulGetFollowers:
      description: The list of followers of the user.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/UserList" }
    SuccessfulFollowUser:
      description: User was added successfully to following.
    SuccessfulUnfollowUser:
      description: User was removed successfully from following.
    SuccessfulBanUser:
      description: User was added successfully to bans.
    SuccessfulUnbanUser:
      description: User was removed successfully from bans.
...
  parameters:
...
    PathUserID:
      description: User identifier passed in path.
      name: userID
      in: path
      schema: { $ref: "#/components/schemas/userID" }
      required: true
    PathPhotoID:
      description: Photo identifier passed in path.
      name: photoID
      in: path
      schema: { $ref: "#/components/schemas/photoID" }
      required: true
...
    StartDate:
      description: |-
        The date of the last record in the previous response.
        It is used for keyset pagination.
        It starts to the current date if not specified.
      name: startDate
      in: query
      schema:
        description: The date of the last record.
        type: string
        format: date-time
        pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
        minLength: 20
        maxLength: 20
      required: false 
    StartId:
      description: |-
        The id of the last record in the previous response.
        It is used for keyset pagination.
        It starts from the first id if it is not specified.
      name: startId
      in: query
      schema:
        description: The id of the last record.
        type: string
        format: uuid
        pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
        minLength: 36
        maxLength: 36
      required: false
...
  securitySchemes:
    bearerAuth:
      description: |-
        Simplified bearer authentication using UUID of the user.
      type: http
      scheme: bearer
      bearerFormat: uuid